split access control tests
